Moved wordlists to a separate module, moved bip39 functions to util.py
