+++ src/seedsigner/helpers/bip39/wordlists/__init__.py
@@ -1,0 +1,5 @@
+from .english_list import WORDLIST__ENGLISH
+from .spanish_list import WORDLIST__SPANISH
+from .french_list import WORDLIST__FRENCH
+from .italian_list import WORDLIST__ITALIAN
+from .portuguese_list import WORDLIST__PORTUGUESE
