+++ src/seedsigner/helpers/bip39/wordlists/english_list.py
