+++ src/seedsigner/helpers/bip39/wordlists/french_list.py
