+++ src/seedsigner/helpers/bip39/wordlists/italian_list.py
