+++ src/seedsigner/helpers/bip39/wordlists/portuguese_list.py
