+++ src/seedsigner/helpers/bip39/wordlists/spanish_list.py
