+++ src/seedsigner/models/decode_qr.py
@@ -15,7 +15,7 @@ from urtypes.bytes import Bytes
 from seedsigner.helpers.ur2.ur_decoder import URDecoder
 from seedsigner.models.qr_type import QRType
 from seedsigner.models.seed import Seed
-from seedsigner.helpers.bip39 import get_bip39_wordlist
+from seedsigner.helpers.bip39.utils import get_bip39_wordlist
 from seedsigner.models.settings import SettingsConstants
